Improved encryption example with Realm Studio reference
