+++ examples/encryptionExample/build.gradle
@@ -6,7 +6,7 @@ android {
     buildToolsVersion rootProject.buildTools
 
     defaultConfig {
-        applicationId 'examples.realm.io.encryptionExample'
+        applicationId 'io.realm.examples.encryption'
         targetSdkVersion rootProject.sdkVersion
         minSdkVersion rootProject.minSdkVersion
         versionCode 1
