+++ examples/encryptionExample/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="io.realm.examples.encryptionexample"
+<manifest package="io.realm.examples.encryption"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
