+++ examples/encryptionExample/src/main/java/io/realm/examples/encryption/Util.java
@@ -13,18 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.realm.examples.encryption;
 
-package io.realm.examples.encryptionexample;
+public class Util {
 
-import android.app.Application;
+    private final static char[] hexArray = "0123456789ABCDEF".toCharArray();
 
-import io.realm.Realm;
-
-public class MyApplication extends Application {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        Realm.init(this);
+    //Original source: https://stackoverflow.com/a/9855338/1389357
+    public static String bytesToHex(byte[] bytes) {
+        char[] hexChars = new char[bytes.length * 2];
+        for ( int j = 0; j < bytes.length; j++ ) {
+            int v = bytes[j] & 0xFF;
+            hexChars[j * 2] = hexArray[v >>> 4];
+            hexChars[j * 2 + 1] = hexArray[v & 0x0F];
+        }
+        return new String(hexChars);
     }
+
 }
